move convertdoitolink fct
